Quit Hatin' on the Forms
The truth is: the form component is super controversial: some people love it, some people hate it - and a lot of people have trouble learning it. In fact, its documentation on symfony.com is read far more than any other section.

Why? Because honestly, it is complex - too complex sometimes. But you know what? The form component is going to allow you to get a lot of work done really quickly. And when I do see someone suffering with forms, most of the time, it's their fault. They create situations that are far more complicated than they need to be.

So let's not do that - let's enjoy forms, and turn them into a weapon.

I made a few changes to the site since last time. For example, go to localhost:8000/genus. It looks the same, but see the 'Sub Family'? Before, that was a string field on Genus. But now, I've added a SubFamily entity:

 40 lines  src/AppBundle/Entity/SubFamily.php
... lines 1 - 6
/**
 * @ORM\Entity
 * @ORM\Table(name='sub_family')
 */
class SubFamily
{
    /**
     * @ORM\Id
     * @ORM\GeneratedValue(strategy='AUTO')
     * @ORM\Column(type='integer')
     */
    private {dollar_sign}id;
    /**
     * @ORM\Column(type='string')
     */
    private ;
... lines 24 - 38
}

And created a ManyToOne relation from Genus to SubFamily:

 119 lines  src/AppBundle/Entity/Genus.php
... lines 1 - 11
class Genus
{
... lines 14 - 25
    /**
     * @ORM\ManyToOne(targetEntity='AppBundle\Entity\SubFamily)
     * @ORM\JoinColumn(nullable=false)
     */
    private {dollar_sign}subFamily;
... lines 31 - 117
}

So every Genus belongs to one SubFamily.

I also started a new admin section - see it at /admin/genus. But, it needs some work - like the ability to add a new genus. That'll be our job. And the code will live in the new GenusAdminController.

Creating a new Form

To create a form, you'll add a class where you'll describe what the form looks like.

 Tip
Actually, you can build a form directly in the controller if you want to.
In PhpStorm, select anywhere in your bundle and press command+N, or right click and select 'New'. Find 'Form' and call it GenusFormType.

This just gave us a basic skeleton and put the class in a Form directory. Sensible! These classes are called 'form types'... which is the worst name that we could come up with when the system was created. Sorry. Really, these classes are 'form recipes'.