Head into the GenusAdminController. I gave us a tiny head start - there's already a newAction().
In the admin area, the 'Add' button points here. Click it!

The controller must return a response (null given). Did you forget to add a return statement somewhere in your controller?
Yep, a great explosion!

Instantiating the Form Object

To create the form object, add  = {dollar_sign}this->createForm(). This is a shortcut method in the base Controller that calls a method on the form.factory service. That's my friendly reminder to you that everything - including form magic - is done by a service.

To createForm(), pass it the form type class name - GenusFormType::class.
And because I used autocomplete, that did just add the use statement for me. The ::class syntax is new in PHP 5.5 - and we're going to use it a lot.