The Form Twig Functions First, we need an opening form tag. Render that with form_start(genusForm). To add a closing form tag, add form_end(genusForm).
I know, having functions to create the HTML form tag seems a little silly. But, wait! form_start() is cool because it will add the enctype=multipart/form-data attribute if the form has an upload field. And the form_end() function takes care of rendering hidden fields. So, these guys are my friends.

Between them, render all three fields at once with form_widget(genusForm).

And finally, we need a button! We can do that by hand: <button type='submit'>, give it a few Bootstrap classes and call it 'Save'

You can also add buttons as fields to your form. But this is helpful in very few cases, so I prefer just to render them by hand.
that's it! Head to the browser and refresh.

There it is! A rendered form with almost no work. They're all text boxes now: we'll customize them soon.

The Bootstrap Form Theme

Of course, it is pretty ugly... Symfony has default HTML markup that it uses to render everything you're seeing: the labels, the inputs and any validation errors.