Inspect the HTML and check out the <form> element. Notice: this does not have an action attribute. This means that the form will submit right back to the same route and controller that renders it. You can totally change this, but we won't.
In other words, our single action method will be responsible for rendering the form and processing it when the request method is POST.

Before we do any processing, we need the request. Type-hint it as an argument.

Next, to actually handle the submit, call {dollar_sign}form->handleRequest() and pass it the {dollar_sign}request object

This is really cool: the  knows what fields it has on it. So {dollar_sign}form->handleRequest() goes out and grabs the post data off of the {dollar_sign}request object for those specific fields and processes them.

The confusing thing is that this only does this for POST requests. If this is a GET request - like the user simply navigated to the form - the handleRequest() method does nothing and our form renders just like it did before.

Tip
You can configure the form to submit on GET requests if you want. This is useful for search forms
But if the form was just submitted, then we'll want to do something with that information, like save a new Genus to the database. Add an if statement: if ({dollar_sign}form->isSubmitted() && {dollar_sign}form->isValid()).
In other words, if this is a POST request and if the form passed all validation. We'll add validation soon.