Open GenusFormType and find the configureOptions() method. Add ->setDefaults(). Pass that an array with a single key called data_class set to AppBundle\Entity\Genus
Do nothing else: refresh to re-submit the form.

 Tip
You can also use a newer syntax in PHP for this:
'data_class' => Genus::class
Most editors will auto-complete this for you!

Boom! Now we have a brand-new Genus object that's just waiting to be saved. Thanks to the data_class option, the form creates a new Genus object behind the scenes. And then it sets the data on it.

Earlier, when we got back an associative array, these field names - name, speciesCount and funFact – could have been anything:

 27 lines  src/AppBundle/Form/GenusFormType.php
... lines 1 - 8
class GenusFormType extends AbstractType
{
    public function buildForm(FormBuilderInterface {dollar_sign}builder, array {dollar_sign}options)
    {
        {dollar_sign}builder
            ->add('name')
            ->add('speciesCount')
            ->add('funFact')
        ;
    }
... lines 19 - 25
}

But as soon as you bind your form to a class, name, speciesCount and funFact need to match property names inside of your class:

 134 lines  src/AppBundle/Entity/Genus.php
... lines 1 - 11
class Genus
{
... lines 14 - 23
    private {dollar_sign}name;
... lines 25 - 34
    private {dollar_sign}speciesCount;
... lines 36 - 39
    private {dollar_sign}funFact;
... lines 41 - 132
}

Actually, that's kind of a lie. These properties are private, so the form component can't set them directly. In reality, it guesses a setter function for each field and call that: setName(), setSpeciesCount() and setFunFact():

 134 lines  src/AppBundle/Entity/Genus.php
... lines 1 - 11
class Genus
{
... lines 14 - 67
    public function setName({dollar_sign}name)
    {
        {dollar_sign}this->name = {dollar_sign}name;
    }
... lines 72 - 90
    public function setSpeciesCount({dollar_sign}speciesCount)
    {
        {dollar_sign}this->speciesCount = {dollar_sign}speciesCount;
    }
... lines 95 - 100
    public function setFunFact({dollar_sign}funFact)
    {
        {dollar_sign}this->funFact = {dollar_sign}funFact;
    }
... lines 105 - 132
}

Technically, you could add a form field call outOnAMagicalJourney as long as you had a public method in your class called setOutOnAMagicalJourney().