Always Redirect! Next, we always redirect after a successful form submit - ya know, to make sure that the user can't just refresh and re-post that data. That'd be lame.
To do that, return ->redirectToRoute(). Hmm, generate a URL to the admin_genus_list route - that's the main admin page I created before the course.

Because redirectToRoute() returns a RedirectResponse, we're done!

Time to try it out. I'll be lazy and refresh the POST. We should get a brand new 'Sea Monster' genus. There it is! Awesome