Adding a Super Friendly (Flash) Message
And I want to build a friendly site, so let's add that message. Back in newAction(), add some code right before the redirect: ->addFlash('success') - you'll see where that key is used in a minute - then Genus created - you are amazing!

It's good to encourage users.

But let's be curious and see what this does behind the scenes. Hold command and click into the addFlash() method:

 398 lines  vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/Controller.php
... lines 1 - 38
abstract class Controller implements ContainerAwareInterface
{
... lines 41 - 102
    /**
     * Adds a flash message to the current session for type.
     *
     * @param string     The type
     * @param string  The message
     *
     * @throws \LogicException
     */
    protected function addFlash({dollar_sign}type, {dollar_sign}message)
    {
            throw new \LogicException('You can not use the addFlash method if sessions are disabled.');
        }
        {dollar_sign}this->container->get('session')->getFlashBag()->add({dollar_sign}type, {dollar_sign}message);
    }
... lines 119 - 396
}

Okay, cool: it uses the session service, fetches something called a 'flash bag' and adds our message to it. So the flash bag is a special part of this session where you can store messages that will automatically disappear after one redirect. If we store a message here and then redirect, on the next page, we can read those messages from the flash bag, and print them on the page. And because the message automatically disappears after one redirect, we won't accidentally show it to the user more than once.

And actually, it's even a little bit cooler than that. A message will actually stay in the flash bag until you ask for it. Then it's removed. This is good because if you - for some reason - redirect twice before rendering the message, no problem! It'll stay in there and wait for you.