So far, we set the property name of the form fields and Symfony tries to guess the best field type to render. Since isPublished is a boolean in the database, it guessed a checkbox field.
But obviously, that's not always going to work.

Google for 'Symfony form field types', and find a document called Form Types Reference(http://symfony.com/doc/current/reference/forms/types.html).

These are all the different field types we can choose from. It's got stuff you'd expect - like text and textarea, some HTML5 fields - like email and integer and some more unusual types, like date, which helps render date fields, either as three drop-downs or a single text field.

Right now, isPublished is a checkbox. But instead, I'd rather have a select drop-down with 'yes' and 'no' options. But... you won't see a 'select' type in this list. Instead, it's called ChoiceType(http://symfony.com/doc/current/reference/forms/types/choice.html).

Using ChoiceType

ChoiceType is a little special: it can render a drop-down, radio buttons or checkboxes based on what options you pass to it. One of the options is choices, which, if you look down at the example, you can see controls the actual items in the drop-down.

Let's use this! In GenusFormType, the optional second argument to the add function is the field type you want to use. Set it to ChoiceType::class.

The third argument is an array of options to configure that field. What can you pass here? Well, each field type has different options... though there are a lot of options shared by all types, like the ability to customize the label.

Either way, the reference section will tell you what you can use. Pass in the choices option and set that to an array. Add Yes mapped to true and No mapped to false. The keys - Yes and No will be the text in the drop down.

The values - true and false will be the value that's passed to setIsPublished() if that option is chosen.

Try it out! Head back to the form and refresh! Perfect.