As cool as it is that it's guessing my field type, I am actually going to add EntityType::class to use this type explicitly.
I don't have to do this: I just want to show you guys what a traditional setup looks like.

Now, do nothing else and refresh. It'll still work, right? Right? Nope!

The required class option is missing! As I just finished saying, we must pass a class option to the EntityType. We got away with this before, because when it's null, Symfony guesses the form 'type' and the class option.