Now, I like to keep all of my queries inside of repository classes because I don't want queries laying around in random places, like in a form class. But, if you look, I don't have a SubFamilyRepository yet.
No worries - copy the GenusRepository, paste it as SubFamilyRepository. Rename that class and clear it out.

Open the SubFamily entity and hook it up with @ORM\Entity(repositoryClass='') and fill in SubFamilyRepository.

Great! Back in our form, we know this repo will be an instance of SubFamilyRepository.

Return {dollar_sign}repo-> and a new method that we're about to create called createAlphabeticalQueryBuilder().

Copy that name and head into the repository to create that function. Inside, return {dollar_sign}this->createQueryBuilder('sub_family') and then order by sub_family.name, ASC.

Done! The query_builder method points here, and we handle the query.

Alright, try it out! Nailed it! As far as form options go, we probably just conquered one of the most complex.