Form Variables
Now notice, most of these functions - including form_row - have a second argument called variables. And judging by this code example, you can apparently control the label with this argument.

Listen closely: these 'variables' are the most powerful part of form rendering. By passing different values, you can override almost every part of how a field is rendered.

So what can you pass here? Scroll down near the bottom to find a big beautiful table called Form Variables Reference(http://symfony.com/doc/current/reference/forms/twig_reference.html#form-variables-reference).

This gives you a big list of all the variables that you can override when rendering a field, including label, attr, label_attr, and other stuff.

To show this off, find the speciesCount field and add a second argument set to {} - the Twig array syntax. Override the label variable: set it to Number of Species.

Refresh! It's just that easy.

The Amazing Form Profiler

There's one huge form tool that we haven't looked at yet. Your web debug toolbar should have a clipboard icon. Click it.

This is the profiler for your form, and it's packed with stuff that's going to make your form life better. If you click speciesCount, you can see the different data for your species - which isn't too interesting on this blank form. You can see any submitted data and all of the variables that can be passed to the field.

It turns out, the reference section we looked at has most of the variables but this will have all of them. This is your place to answer:

What are the values for my variables? And, what can I override?
This also shows you 'Resolved Options': these are the final values of the options that can be passed as the third argument to the add() function.

CSRF Protection

Oh, and check out this _token field. Do you remember adding this?

Hopefully not - because we never did! This is a CSRF token that's automatically added to the form. It's rendered for us when we call form_end and validated behind the scenes along with all the other fields. It's free CSRF protection.