Check out the web debug toolbar: it's highlighted with the number of validation errors, which is pretty cool. It's even cooler because we can see where those are coming from.
The default message for the NotBlank constraint is obviously

This value should not be blank
We can easily change this by passing a message option to the annotation. But wait a second: we're going to use NotBlank a lot. Is there a way to customize the default message across the whole system?

Yea! All of these strings are passed through Symfony's translator. And we can take advantage of that to customize this message, even in English.

Enabling the Translator

First, in case you don't already have it enabled, open up app/config/config.yml. Activate the translator service by uncommenting out the translator key under framework.

Refresh this page and watch the web debug toolbar. Suddenly, there's an extra icon that's coming from the translator. It's reporting that there are ten missing messages. In other words, we are apparently already sending ten messages through the translator that are missing translation strings.