Changing Validation Messages
But, if you want to customize this message, copy it. And, notice, the domain is validators: that's basically a translation 'category', and it's important for what we do next.

We don't have any translation files yet, so create a new directory called translations in app/Resources. Inside, add a new file: validators.en.yml. This is validators because the message is being translated in that domain.

Inside, paste the string and set it to 'Hi! Please enter something for this field.'.

And that's it! Go back and refresh! Oh no, it didn't work! Well, I kind of expected that. Find your terminal, open a new tab, and run:

bin/console cache:clear
You almost never need to worry about clearing your cache while developing but very occasionally, you'll find a quirk in Symfony that needs this. Sometimes, when you add a new translation file, this happens.

Let's refresh again. There it is!

So yay validation! There's one more constraint I want you to see: Callback. This is your Swiss army knife: it allows you to write whatever custom validation logic you want inside a method. There, you can create different validation errors and map them to any field.