Passing in Default Data This form needs to be pre-filled with all of the data from the database. So again, how can I pass default data to a form? It's as simple as this: the second argument to createForm is the default data. Pass it the entire  object.
Why the entire object? Because remember: our form is bound to the Genus class:
50 lines  src/AppBundle/Form/GenusFormType.php
... lines 1 - 13
class GenusFormType extends AbstractType
{
... lines 16 - 42
    public function configureOptions(OptionsResolver {dollar_sign}resolver)
    {
        {dollar_sign}resolver->setDefaults([
            'data_class' => 'AppBundle\Entity\Genus'
        ]);
    }
}

That means that its output will be a Genus object, but its input should also be a Genus object.

Behind the scenes, it will use the getter functions on Genus to pre-fill the form: like getName(). And everything else is exactly the same. Well, I'll tweak the flash message but you get the idea.